+++ src/main.rs
@@ -79,7 +79,7 @@ async fn entrypoint(globals: Arc<Globals>, backends: Arc<HashMap<String, Backend
   let futures = select_all(addresses.into_iter().map(|addr| {
     let mut tls_enabled = false;
     if let Some(https_port) = globals.https_port {
-      tls_enabled = https_port == (addr.port() as u32)
+      tls_enabled = https_port == (addr.port() as u16)
     }
 
     info!("Listen address: {:?} (TLS = {})", addr, tls_enabled);
